Issue 4513 move classes
