+++ api/src/main/java/com/epam/pipeline/manager/cluster/performancemonitoring/monitor/common/RunMonitorUtils.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.epam.pipeline.manager.cluster.performancemonitoring.monitor;
+package com.epam.pipeline.manager.cluster.performancemonitoring.monitor.common;
 
 import com.epam.pipeline.common.MessageConstants;
 import com.epam.pipeline.common.MessageHelper;
